Continue MinGW-64 updating work in core PCP - libpcp builds
Small stuff, all over the map a bit:
- additional guards to the use of Unix domain sockets;
- remove manual errno mappings now in toolchain headers;
- re-order the header includes to match toolchain requirements;
- provide a simple strcasestr implementation for getdate code;
- provide simple inet_ntop/pton implementations for socket code;
- drop old TIMEOUT_ASYNC macro - conflicts with system headers;
- rename local variable 'interface' in auxserver, compile fail;
- drop use of a uint16_t cast, type unavailable and not needed;
- conditional compilation guards around getuid/getgid usage.

Updated Makepkgs builds for Windows via the fedora-mingw project
mechanisms.

+++ src/libpcp/src/discovery.c
@@ -127,9 +127,9 @@ __pmAddDiscoveredService(__pmServiceInfo *info, int numUrls, char ***urls)
 		  size, PM_FATAL_ERR);
     }
     if (isIPv6)
-	snprintf(url, size, "%s://[%s]:%u", protocol, address, (uint16_t)port);
+	snprintf(url, size, "%s://[%s]:%u", protocol, address, port);
     else
-	snprintf(url, size, "%s://%s:%u", protocol, address, (uint16_t)port);
+	snprintf(url, size, "%s://%s:%u", protocol, address, port);
     free(address);
 
     /*
